One exception per class
