+++ Microsoft.NET.Build.Containers/ContainerHttpException.cs
@@ -1,0 +1,13 @@
+namespace Microsoft.NET.Build.Containers;
+public class ContainerHttpException : Exception
+{
+    private const string errorPrefix = "Containerize: error CONTAINER004:";
+    string? jsonResponse;
+    string? uri;
+    public ContainerHttpException(string message, string? targetUri, string? jsonResp)
+            : base($"{errorPrefix} {message}\nURI: {targetUri ?? "None."}\nJson Response: {jsonResp ?? "None."}")
+    {
+        jsonResponse = jsonResp;
+        uri = targetUri;
+    }
+}
