+++ Microsoft.NET.Build.Containers/DockerLoadException.cs
@@ -20,16 +20,3 @@ public class DockerLoadException : Exception
     {
     }
 }
-
-public class ContainerHttpException : Exception
-{
-    private const string errorPrefix = "Containerize: error CONTAINER004:";
-    string? jsonResponse;
-    string? uri;
-    public ContainerHttpException(string message, string? targetUri, string? jsonResp) 
-            : base($"{errorPrefix} {message}\nURI: {targetUri ?? "None."}\nJson Response: {jsonResp ?? "None."}" )
-    {
-        jsonResponse = jsonResp;
-        uri = targetUri;
-    }
-}
